refactor EventToolkit
