+++ taichi/program/kernel_profiler.h
@@ -103,6 +103,45 @@ class KernelProfilerBase {
   }
 };
 
+class EventToolkitBase {
+ public:
+  virtual void update_record(uint32_t records_size_after_sync,
+                     std::vector<KernelProfileTracedRecord> &traced_records) {
+                      TI_NOT_IMPLEMENTED;
+                     };
+  virtual KernelProfilerBase::TaskHandle start_with_handle(
+      const std::string &kernel_name) {
+        TI_NOT_IMPLEMENTED;
+      };
+  virtual void update_timeline(std::vector<KernelProfileTracedRecord> &traced_records) {
+    TI_NOT_IMPLEMENTED;
+  };
+
+ private:
+  struct EventRecord {
+    std::string name;
+    float kernel_elapsed_time_in_ms{0.0};
+    float time_since_base{0.0};
+    void *start_event{nullptr};
+    void *stop_event{nullptr};
+  };
+  float64 base_time_{0.0};
+  void *base_event_{nullptr};
+  // for cuEvent profiling, clear after sync()
+  std::vector<EventRecord> event_records_;
+
+ public:
+  void clear() {
+    event_records_.clear();
+  }
+  EventRecord *get_current_event_record() {
+    return &(event_records_.back());
+  }
+  void *get_base_event() const {
+    return base_event_;
+  }
+};
+
 std::unique_ptr<KernelProfilerBase> make_profiler(Arch arch, bool enable);
 
 }  // namespace taichi::lang
