+++ taichi/rhi/amdgpu/amdgpu_profiler.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "taichi/system/timeline.h"
+#include "taichi/program/kernel_profiler.h"
+#include "taichi/rhi/amdgpu/amdgpu_driver.h"
+#include "taichi/rhi/amdgpu/amdgpu_context.h"
+
+#include <string>
+#include <stdint.h>
+
+namespace taichi::lang {
+    
+}
