+++ taichi/rhi/cuda/cuda_profiler.h
@@ -17,7 +17,7 @@ enum class ProfilingToolkit : int {
   cupti,
 };
 
-class EventToolkit;
+class EventToolkitCUDA;
 
 // A CUDA kernel profiler
 class KernelProfilerCUDA : public KernelProfilerBase {
@@ -55,44 +55,20 @@ class KernelProfilerCUDA : public KernelProfilerBase {
 
   // Instances of these toolkits may exist at the same time,
   // but only one will be enabled.
-  std::unique_ptr<EventToolkit> event_toolkit_{nullptr};
+  std::unique_ptr<EventToolkitCUDA> event_toolkit_{nullptr};
   std::unique_ptr<CuptiToolkit> cupti_toolkit_{nullptr};
   std::vector<std::string> metric_list_;
   uint32_t records_size_after_sync_{0};
 };
 
 // default profiling toolkit
-class EventToolkit {
+class EventToolkitCUDA : public EventToolkitBase {
  public:
   void update_record(uint32_t records_size_after_sync,
-                     std::vector<KernelProfileTracedRecord> &traced_records);
+                     std::vector<KernelProfileTracedRecord> &traced_records) override;
   KernelProfilerBase::TaskHandle start_with_handle(
-      const std::string &kernel_name);
-  void update_timeline(std::vector<KernelProfileTracedRecord> &traced_records);
-  void clear() {
-    event_records_.clear();
-  }
-
- private:
-  struct EventRecord {
-    std::string name;
-    float kernel_elapsed_time_in_ms{0.0};
-    float time_since_base{0.0};
-    void *start_event{nullptr};
-    void *stop_event{nullptr};
-  };
-  float64 base_time_{0.0};
-  void *base_event_{nullptr};
-  // for cuEvent profiling, clear after sync()
-  std::vector<EventRecord> event_records_;
-
- public:
-  EventRecord *get_current_event_record() {
-    return &(event_records_.back());
-  }
-  void *get_base_event() const {
-    return base_event_;
-  }
+      const std::string &kernel_name) override;
+  void update_timeline(std::vector<KernelProfileTracedRecord> &traced_records) override;
 };
 
 }  // namespace taichi::lang
